DataLayer: replace span with context

+++ pkg/persistence/helm.go
@@ -8,25 +8,31 @@ import (
 	"github.com/dollarshaveclub/acyl/pkg/models"
 	"github.com/lib/pq"
 	"github.com/pkg/errors"
-	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
 )
 
 var _ HelmDataLayer = &PGLayer{}
 
-func (pg *PGLayer) GetHelmReleasesForEnv(span tracer.Span, name string) ([]models.HelmRelease, error) {
+func (pg *PGLayer) GetHelmReleasesForEnv(ctx context.Context, name string) ([]models.HelmRelease, error) {
+	if isCancelled(ctx.Done()) {
+		return nil, errors.Wrap(ctx.Err(), "error getting helm releases for env")
+	}
 	q := `SELECT ` + models.HelmRelease{}.Columns() + ` FROM helm_releases WHERE env_name = $1;`
-	ctx := tracer.ContextWithSpan(context.Background(), span)
 	return collectHelmRows(pg.db.QueryContext(ctx, q, name))
 }
 
-func (pg *PGLayer) UpdateHelmReleaseRevision(span tracer.Span, envname, release, revision string) error {
+func (pg *PGLayer) UpdateHelmReleaseRevision(ctx context.Context, envname, release, revision string) error {
+	if isCancelled(ctx.Done()) {
+		return errors.Wrap(ctx.Err(), "error updating helm relese revision")
+	}
 	q := `UPDATE helm_releases SET revision_sha = $1 WHERE env_name = $2 AND release = $3;`
-	ctx := tracer.ContextWithSpan(context.Background(), span)
 	_, err := pg.db.ExecContext(ctx, q, revision, envname, release)
 	return errors.Wrap(err, "error updating helm release")
 }
 
-func (pg *PGLayer) CreateHelmReleasesForEnv(span tracer.Span, releases []models.HelmRelease) error {
+func (pg *PGLayer) CreateHelmReleasesForEnv(ctx context.Context, releases []models.HelmRelease) error {
+	if isCancelled(ctx.Done()) {
+		return errors.Wrap(ctx.Err(), "error creating helm releases for env")
+	}
 	tx, err := pg.db.Begin()
 	if err != nil {
 		return errors.Wrap(err, "error opening txn")
@@ -50,9 +56,11 @@ func (pg *PGLayer) CreateHelmReleasesForEnv(span tracer.Span, releases []models.
 	return nil
 }
 
-func (pg *PGLayer) DeleteHelmReleasesForEnv(span tracer.Span, name string) (uint, error) {
+func (pg *PGLayer) DeleteHelmReleasesForEnv(ctx context.Context, name string) (uint, error) {
+	if isCancelled(ctx.Done()) {
+		return 0, errors.Wrap(ctx.Err(), "error deleting helm releases for env")
+	}
 	q := `DELETE FROM helm_releases WHERE env_name = $1;`
-	ctx := tracer.ContextWithSpan(context.Background(), span)
 	res, err := pg.db.ExecContext(ctx, q, name)
 	n, _ := res.RowsAffected()
 	return uint(n), err
